Fix plugins not loading with custom home directory
If Jenkins uses a non-default home directory, plugins will not be loaded
because they are installed in a hard-coded location. This adds the
`plugin_parent_dir` option to `Jenkins` class, which allows specifying
where to install plugins.

If `plugin_parent_dir` is not specified, the previously hard-coded path
('/var/lib/jenkins') is used.